move standalone to executors module
