+++ tests/test_web/test_module.py
@@ -4,7 +4,7 @@ import pytest
 from servestatic import ServeStaticASGI
 
 import reactpy
-from reactpy.asgi.standalone import ReactPy
+from reactpy.asgi.executors.standalone import ReactPy
 from reactpy.testing import (
     BackendFixture,
     DisplayFixture,
